requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
# AWS Lambda adapter
mangum>=0.17.0
# JWT validation for Cognito
python-jose[cryptography]>=3.3.0
anthropic>=0.40.0
openai>=1.0.0
aiofiles>=23.2.0
# LangChain/LangSmith dependencies
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langchain-voyageai>=0.2.0
langchain-chroma>=0.1.0
langchain-postgres>=0.0.9
langsmith>=0.1.100
langchainhub>=0.1.14
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0
tavily-python>=0.3.0
# LangChain community and text processing
chromadb>=0.4.0
rank-bm25>=0.2.0
langchain-text-splitters>=0.3.0
# Retrieval system dependencies
rouge>=1.0.1
pandas>=2.0.0
# Vector search dependencies
voyageai>=0.2.0
numpy>=1.24.0
# Golden dataset tool dependencies
scikit-learn>=1.3.0
# Evaluation script dependencies
requests>=2.31.0
# PostgreSQL and SQLAlchemy dependencies
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
boto3>=1.34.0
# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
moto>=5.0.0
